feat(task): promote task schema to nab::task library module (MIK-5359)
Slice-4 step 1 (the binary-boundary prerequisite, design §12.2): move the
task-engine schema — TaskAction, TaskStatus, DiscoveredApi (+From<ApiEndpoint>),
TaskOutcome, ActionObservation — from the binary-only src/cmd/task.rs to a new
nab::task LIBRARY module (src/task.rs), behind the task feature.

Why: the slice-4 self-contained sampling loop lives in nab-mcp (only it can
call sampling/createMessage) and cannot reach cmd::task across the binary
boundary. Putting the schema in the library lets both consumers — the nab CLI
executor (cmd::task) and the nab-mcp loop — share one contract instead of
duplicating it. cmd::task now imports from nab::task and keeps the executor
(execute_action / discover_apis / fetch_config_for_api_call / cmd_task).

As pub library API the types are no longer dead-code, so the #[allow(dead_code)]
allows are dropped. The 5 pure schema tests move to the lib module; the 5
executor tests stay in cmd. CI's task-feature job updated to
`--lib --bin nab task` so it runs BOTH (the old `--bin nab cmd::task` would
have missed the moved lib tests).

cmd_fetch untouched. Feature stays out of default.

Verified under CI's gate: fmt clean; default RUSTFLAGS=-Dwarnings build clean
(task module gates out, no dead-code leak); clippy --features task -D warnings
clean; `cargo test --features task --lib --bin nab task` => lib 5 + bin 5 passed.

Local-Tested: cargo fmt+clippy+test green @ 8a5e411eafc18ee9a7e7f250ac5c3f81a3b19166
